requirements.txt
mysql-connector-python~=9.3.0
Flask~=3.1.0
Flask-Login~=0.6.3
Werkzeug~=3.1.3